Renamed `Instantiating` namespace to `PresenterFactory`
